ref: move error classes into logging.ts
