--- src/utils/errors.ts
@@ -1,14 +1,0 @@
-
-export class LyricsError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "LyricsError";
-  }
-}
-
-export class PluginError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "PluginError";
-  }
-}
