+++ src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from "./dom.js";
-export * from "./errors.js";
 export * from "./input.js";
 export * from "./logging.js";
 export * from "./misc.js";
